Fix C++ tests build, move test to test_query_builder.py
