+++ python/tests/unit/arcticdb/version_store/test_filtering.py
@@ -2314,24 +2314,3 @@ def test_filter_with_column_slicing_defragmented(lmdb_version_store_tiny_segment
         received = lmdb_version_store_tiny_segment.read(symbol, query_builder=q).data
         expected = df.query(pandas_query)
         assert np.array_equal(expected, received) and (not expected.empty and not received.empty)
-
-
-def test_filter_sparse(lmdb_version_store):
-    # Remove this test and replace with more extensive tests once this issue is fixed:
-    # https://github.com/man-group/ArcticDB/issues/1404
-    lib = lmdb_version_store
-    sym = "test_filter_sparse"
-    df = pd.DataFrame({"col": [0.0, np.nan, 0.0]}, index=pd.date_range("2024-01-01", periods=3))
-    lib.write(sym, df, sparsify_floats=True)
-
-    # These 2 queries exercise different code paths
-    q = QueryBuilder()
-    q = q[q["col"].isnull()]
-    with pytest.raises(InternalException):
-        lib.read(sym, query_builder=q)
-
-    q = QueryBuilder()
-    q = q[q["col"] == np.float64(0.0)]
-    with pytest.raises(InternalException):
-        lib.read(sym, query_builder=q)
-
